Rearranged the project for a cleaner file structure better for understanding
